PR #64 - refactor connection logic out to a Connectable mixin so TrustedAdvisor doesn't have to be a subclass of _AwsService
