+++ awslimitchecker/tests/services/test_base.py
@@ -272,80 +272,6 @@ class Test_AwsService(object):
         assert res == {'foo': mock_limit1, 'foo4': mock_limit4}
         assert mock_find_usage.mock_calls == [call()]
 
-    def test_connect_via_no_region(self):
-        cls = AwsServiceTester(1, 2)
-        mock_driver = Mock()
-        res = cls.connect_via(mock_driver)
-        assert mock_driver.mock_calls == [
-            call(None)
-        ]
-        assert res == mock_driver.return_value
-
-    def test_connect_via_with_region(self):
-        cls = AwsServiceTester(1, 2, region='foo')
-        mock_driver = Mock()
-        with patch('awslimitchecker.services.base._AwsService.'
-                   '_get_sts_token') as mock_get_sts:
-            res = cls.connect_via(mock_driver)
-        assert mock_get_sts.mock_calls == []
-        assert mock_driver.mock_calls == [
-            call('foo')
-        ]
-        assert res == mock_driver.return_value
-
-    def test_connect_via_sts(self):
-        cls = AwsServiceTester(1, 2, account_id='123', account_role='myrole',
-                               region='myregion')
-        mock_driver = Mock()
-        mock_creds = Mock()
-        type(mock_creds).access_key = 'sts_ak'
-        type(mock_creds).secret_key = 'sts_sk'
-        type(mock_creds).session_token = 'sts_token'
-
-        with patch('awslimitchecker.services.base._AwsService.'
-                   '_get_sts_token') as mock_get_sts:
-            mock_get_sts.return_value = mock_creds
-            res = cls.connect_via(mock_driver)
-        assert mock_get_sts.mock_calls == [call()]
-        assert mock_driver.mock_calls == [
-            call(
-                'myregion',
-                aws_access_key_id='sts_ak',
-                aws_secret_access_key='sts_sk',
-                security_token='sts_token'
-            )
-        ]
-        assert res == mock_driver.return_value
-
-    def test_get_sts_token(self):
-        cls = AwsServiceTester(1, 2, account_id='789',
-                               account_role='myr', region='foobar')
-        with patch('awslimitchecker.services.base.boto.sts.connect_to_region'
-                   '') as mock_connect:
-            res = cls._get_sts_token()
-        arn = 'arn:aws:iam::789:role/myr'
-        assert mock_connect.mock_calls == [
-            call('foobar'),
-            call().assume_role(arn, 'awslimitchecker', external_id=None),
-        ]
-        assume_role_ret = mock_connect.return_value.assume_role.return_value
-        assert res == assume_role_ret.credentials
-
-    def test_get_sts_token_external_id(self):
-        cls = AwsServiceTester(1, 2, account_id='789',
-                               account_role='myr', region='foobar',
-                               external_id='myextid')
-        with patch('awslimitchecker.services.base.boto.sts.connect_to_region'
-                   '') as mock_connect:
-            res = cls._get_sts_token()
-        arn = 'arn:aws:iam::789:role/myr'
-        assert mock_connect.mock_calls == [
-            call('foobar'),
-            call().assume_role(arn, 'awslimitchecker', external_id='myextid'),
-        ]
-        assume_role_ret = mock_connect.return_value.assume_role.return_value
-        assert res == assume_role_ret.credentials
-
 
 class Test_AwsServiceSubclasses(object):
